Renamed files & refactor for new class names
