+++ src/include/optimizer/property_enforcer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "optimizer/group_expression.h"
-#include "optimizer/operator_expression.h"
+#include "optimizer/operator_node.h"
 #include "optimizer/property_visitor.h"
 
 namespace terrier::optimizer {
