+++ test/optimizer/physical_operator_test.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 #include "catalog/catalog_defs.h"
 #include "common/managed_pointer.h"
-#include "optimizer/operator_expression.h"
+#include "optimizer/operator_node.h"
 #include "optimizer/physical_operators.h"
 #include "parser/expression/abstract_expression.h"
 #include "parser/expression/constant_value_expression.h"
